backend/StudHome/requirements.txt
 
djangorestframework
environs
django-phonenumber-field==8.1.0
psycopg2-binary==2.9.10
django-cloudinary-storage==0.3.0
pillow==11.3.0
cloudinary==1.44.1
python-decouple==3.8